temp EmailSection refactor
